Add shared API definition for response
This is a bit of a journey / compromise. Specifically, just by creating a `d.ts`
file that we want to share, we've kinda broken into 'this stuff is broken' jail.

1. We want to use single outfile for typescript.
2. We're (currently) lazy, and not using a bundler
3. Imports can't *really* be used with single file output

So you create the `d.ts`, but to use it like an import from the server, it needs
exports. But see point three above. Grrr.

Giving up on the the 'single file' output for the moment lets us kinda get away
with it for the short term -- our imports here are just types, so get tossed out
during compilation.

The API structure here probably won't hold, but it let us shake the tree.