Handle core tweet types & render them
Now all four core types of tweet are displayed & render their type.
The `TweetType` was removed in favour of discriminated unions.

Note the "hack" in the service casting the tweet type to `any`, because
of the way the types are currenty declared.

Placeholder styles also created to differentiate quoted tweets