Complete support of RT'd QT's
Retweet's of Quote tweets look like quote tweets, but aren't quite the same. The
detection logic we had in place was confused (hence earlier commit to fix the
crash). It needed a revamp.

This change tweaks this by removing the discrete case of 'a retweet', and
considers retweeting an attribute of a tweet. While working on this change its
become clear that more work is needed, but for now this solves the problem.

Additionally, a sample of the 'weird' case is added and included in the canned
response.