Add in-timeline replies support
We will no seek & bundle up any replies into threads that are hung off the most
recent reply in the thread. This makes assumptions about the order of the
response, so required some tweaks to the randomness of the replies --
specifically, forcing the earlier parts of the thread to the very end of the
list, after shuffling.

Additionally, Changed type definition of twitters tweets to make `stiched.ts`
show fewer errors