Support replied-to tweets that aren't in the timeline
If a reply is to a tweet that isn't in the main timeline response (too old, for
example), we need to retreive it.

With this change we can now do that -- after we've spelunked through the main
timeline, we will use any threads that still contain a 'reply to' ID that isn't
seen to bulk request replies. This will keep going until they're found, so may
take a moment to do it's thing.

Additionally, so we can iterate over `Set<T>` and `Map<T>` updated the target ES
runtime to `ES2020`